move util fxns into blokus.py
